moved code to cpp and added thread local plugin, changed method of calling the objects member functions to PyObject_CallMethodObjArgs

+++ src/scale.py
@@ -28,7 +28,7 @@ class plugin(Plugin):
 
     def augment_audio(self, mat):
         if self.do_scale:
-            mat2 = (mat * self.amplitude).astype(np.int16)
+            mat2 = (mat.astype(np.float32) * self.amplitude).astype(np.int16)
             return mat2
         else:
             dst = mat
